bugfix for genre column

+++ db/migrations/20190726091933_addGenre.js
@@ -4,7 +4,9 @@ exports.up = function(knex, Promise) {
   };
   
   exports.down = function(knex, Promise) {
-    return knex.schema.table('books', booksTable => booksTable.dropColumn('genre'))
+    if (knex.schema.hasColumn('books', 'genre')) {
+      return knex.schema.table('books', booksTable => booksTable.dropColumn('genre'))
+    }
   };
